requirements.txt
# Core RAG Framework
langchain==0.1.0
chromadb==0.4.18
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pillow==10.1.0
opencv-python==4.8.1.78

# Computer Vision & NLP
transformers==4.36.2
torch==2.1.1
torchvision==0.16.1
accelerate==0.25.0

# Web Interface
streamlit==1.29.0
streamlit-chat==0.1.1
streamlit-extras==0.3.5

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0